@@ -6,19 +6,29 @@ import { ExerciseTopicService } from 'src/app/shared/services/exercise-topic.ser
 import { Course } from 'src/app/shared/models/course.model';
 import { CourseService } from 'src/app/shared/services/course.service';
 import { CourseType } from 'src/app/shared/utils/course-type.model';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, Validators, ReactiveFormsModule } from '@angular/forms';
 import { ExercisePostParameters } from 'src/app/shared/models/exercise-post-parameters.model';
+import { MainButtonComponent } from '../../shared/components/main-button/main-button.component';
+import { NgFor } from '@angular/common';
+import { RatingModule } from 'ngx-bootstrap/rating';
 
 @Component({
-  selector: 'app-exercise-form',
-  templateUrl: './exercise-form.component.html',
-  styleUrls: ['./exercise-form.component.scss'],
+    selector: 'app-exercise-form',
+    templateUrl: './exercise-form.component.html',
+    styleUrls: ['./exercise-form.component.scss'],
+    standalone: true,
+    imports: [
+        ReactiveFormsModule,
+        RatingModule,
+        NgFor,
+        MainButtonComponent,
+    ],
 })
 export class AddExerciceComponent implements OnInit {
   @Input() baseExercise?: Exercise;
   @Output() onValidate = new EventEmitter<void>();
 
-  form!: FormGroup;
+  form!: UntypedFormGroup;
 
   courseList: Course[] = [];
   courseListForSelectedType: Course[] = [];
@@ -31,7 +41,7 @@ export class AddExerciceComponent implements OnInit {
     private exerciseService: ExerciseService,
     private courseService: CourseService,
     private exerciseTopicService: ExerciseTopicService,
-    private formBuilder: FormBuilder
+    private formBuilder: UntypedFormBuilder
   ) {}
 
   ngOnInit(): void {
