@@ -2,35 +2,37 @@ import { Component, OnInit } from '@angular/core';
 import { AuthService } from '../../services/auth.service';
 import { BsModalRef } from 'ngx-bootstrap/modal';
 import { User } from '../../models/user.model';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, Validators, ReactiveFormsModule } from '@angular/forms';
 import { Router } from '@angular/router';
+import { NgIf } from '@angular/common';
+import { MainButtonComponent } from '../main-button/main-button.component';
 
 @Component({
-  selector: 'app-account-popup',
-  templateUrl: './account-popup.component.html',
-  styleUrls: ['./account-popup.component.scss'],
+    selector: 'app-account-popup',
+    templateUrl: './account-popup.component.html',
+    styleUrls: ['./account-popup.component.scss'],
+    standalone: true,
+    imports: [
+        MainButtonComponent,
+        NgIf,
+        ReactiveFormsModule,
+    ],
 })
 export class AccountPopupComponent implements OnInit {
   loggedUser!: User;
 
   editPassword: boolean = false;
 
-  passwordForm!: FormGroup;
+  passwordForm!: UntypedFormGroup;
 
   constructor(
     private bsModalRef: BsModalRef,
     private authService: AuthService,
-    private formBuilder: FormBuilder,
+    private formBuilder: UntypedFormBuilder,
     private router: Router
   ) {}
 
   ngOnInit(): void {
-    if (this.bsModalRef.content) {
-      this.loggedUser = this.bsModalRef.content.loggedUser;
-    } else {
-      this.bsModalRef.hide();
-    }
-
     this.passwordForm = this.formBuilder.group(
       {
         oldPassword: [
@@ -62,7 +64,7 @@ export class AccountPopupComponent implements OnInit {
     );
   }
 
-  checkPasswords(group: FormGroup) {
+  checkPasswords(group: UntypedFormGroup) {
     if (
       group.get('newPassword')?.value !==
       group.get('newPasswordConfirmation')?.value
@@ -76,7 +78,7 @@ export class AccountPopupComponent implements OnInit {
   logout() {
     this.authService.logout();
     this.bsModalRef.hide();
-    this.router.navigate(['/acceuil']);
+    this.router.navigate(['/accueil']);
   }
 
   togglePasswordEdition() {
