@@ -1,22 +1,32 @@
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, Validators, ReactiveFormsModule } from '@angular/forms';
 import { AuthService } from '../../services/auth.service';
 import { of, switchMap } from 'rxjs';
 import { BsModalRef } from 'ngx-bootstrap/modal';
+import { MainButtonComponent } from '../main-button/main-button.component';
+import { NgIf } from '@angular/common';
+import { BackgroundCardComponent } from '../background-card/background-card.component';
 
 @Component({
-  selector: 'app-login-popup',
-  templateUrl: './login-popup.component.html',
-  styleUrls: ['./login-popup.component.scss'],
+    selector: 'app-login-popup',
+    templateUrl: './login-popup.component.html',
+    styleUrls: ['./login-popup.component.scss'],
+    standalone: true,
+    imports: [
+        BackgroundCardComponent,
+        ReactiveFormsModule,
+        NgIf,
+        MainButtonComponent,
+    ],
 })
 export class LoginPopupComponent implements OnInit {
-  loginForm!: FormGroup;
-  registerForm!: FormGroup;
+  loginForm!: UntypedFormGroup;
+  registerForm!: UntypedFormGroup;
 
   constructor(
     private bsModalRef: BsModalRef,
     private authService: AuthService,
-    private formBuilder: FormBuilder
+    private formBuilder: UntypedFormBuilder
   ) {}
 
   ngOnInit(): void {
@@ -73,7 +83,7 @@ export class LoginPopupComponent implements OnInit {
     }
   }
 
-  checkPasswords(group: FormGroup) {
+  checkPasswords(group: UntypedFormGroup) {
     if (
       group.get('password')?.value !==
       group.get('passwordConfirmation')?.value
