@@ -1,26 +1,23 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 import { BsModalRef } from 'ngx-bootstrap/modal';
 import { Charbon } from 'src/app/shared/models/charbon.model';
 import { CharbonService } from 'src/app/shared/services/charbon.service';
+import { AddCharbonComponent } from '../charbon-form/charbon-form.component';
 
 @Component({
-  selector: 'app-edit-popup',
-  templateUrl: './charbon-edition-popup.component.html',
-  styleUrls: ['./charbon-edition-popup.component.scss'],
+    selector: 'app-edit-popup',
+    templateUrl: './charbon-edition-popup.component.html',
+    styleUrls: ['./charbon-edition-popup.component.scss'],
+    standalone: true,
+    imports: [AddCharbonComponent],
 })
-export class CharbonEditionPopup implements OnInit {
+export class CharbonEditionPopup {
   editedCharbon!: Charbon;
 
   constructor(
     private bsModalRef: BsModalRef,
     private charbonService: CharbonService
   ) {}
-
-  ngOnInit(): void {
-    if (this.bsModalRef.content) {
-      this.editedCharbon = this.bsModalRef.content.editedCharbon;
-    }
-  }
 
   close(): void {
     this.bsModalRef.hide();
