@@ -3,22 +3,19 @@ import { BsModalRef } from 'ngx-bootstrap/modal';
 import { ExerciseTopic } from 'src/app/shared/models/exercise-topic.model';
 import { Exercise } from 'src/app/shared/models/exercise.model';
 import { ExerciseService } from 'src/app/shared/services/exercise.service';
+import { AddExerciceComponent } from '../exercise-form/exercise-form.component';
 
 @Component({
-  selector: 'app-exercise-edition-popup',
-  templateUrl: './exercise-edition-popup.component.html',
-  styleUrls: ['./exercise-edition-popup.component.scss']
+    selector: 'app-exercise-edition-popup',
+    templateUrl: './exercise-edition-popup.component.html',
+    styleUrls: ['./exercise-edition-popup.component.scss'],
+    standalone: true,
+    imports: [AddExerciceComponent]
 })
-export class ExerciseEditionPopupComponent implements OnInit {
+export class ExerciseEditionPopupComponent {
   editedExercise!: Exercise;
 
   constructor(private bsModalRef: BsModalRef, private exerciseService: ExerciseService) {}
-
-  ngOnInit(): void {
-    if (this.bsModalRef.content) {
-      this.editedExercise = this.bsModalRef.content.editedExercise;
-    }
-  }
 
   close(): void {
     this.bsModalRef.hide();
