@@ -2,11 +2,27 @@ import { Component, OnInit } from '@angular/core';
 import { AnnouncementGetParameters } from 'src/app/shared/models/announcement-get-parameters.model';
 import { Announcement } from 'src/app/shared/models/announcement.model';
 import { AnnouncementService } from 'src/app/shared/services/announcement.service';
+import { AnnouncementCardComponent } from '../announcement-card/announcement-card.component';
+import { NgFor } from '@angular/common';
+import { BackgroundCardComponent } from '../../shared/components/background-card/background-card.component';
+import { ReactiveFormsModule, FormsModule } from '@angular/forms';
+import { RouterLink } from '@angular/router';
+import { MainButtonComponent } from '../../shared/components/main-button/main-button.component';
 
 @Component({
-  selector: 'app-announcements-page',
-  templateUrl: './announcements-page.component.html',
-  styleUrls: ['./announcements-page.component.scss'],
+    selector: 'app-announcements-page',
+    templateUrl: './announcements-page.component.html',
+    styleUrls: ['./announcements-page.component.scss'],
+    standalone: true,
+    imports: [
+        MainButtonComponent,
+        RouterLink,
+        ReactiveFormsModule,
+        FormsModule,
+        BackgroundCardComponent,
+        NgFor,
+        AnnouncementCardComponent,
+    ],
 })
 export class AnnouncementsPageComponent implements OnInit {
   announcementList: Announcement[] = [];
