@@ -1,9 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 
 @Component({
-  selector: 'app-social-networks',
-  templateUrl: './social-networks.component.html',
-  styleUrls: ['./social-networks.component.scss']
+    selector: 'app-social-networks',
+    templateUrl: './social-networks.component.html',
+    styleUrls: ['./social-networks.component.scss'],
+    standalone: true
 })
 export class SocialNetworksComponent implements OnInit {
 
